# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/vehicles": {
      "get": {
        "operationId": "FleetController_listVehicles",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Vehículos del alcance con elegibilidad, condiciones y lifecycle calculados por el servidor.",
        "tags": [
          "fleet"
        ]
      },
      "post": {
        "operationId": "FleetController_createVehicle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "fleet"
        ]
      }
    },
    "/api/v1/vehicles/{id}": {
      "patch": {
        "operationId": "FleetController_updateVehicle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Actualización parcial con versión optimista (409 si cambió). Odómetro nunca retrocede (EXC-003).",
        "tags": [
          "fleet"
        ]
      }
    },
    "/api/v1/vehicles/{id}/block": {
      "post": {
        "operationId": "FleetController_block",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "fleet"
        ]
      }
    },
    "/api/v1/vehicles/{id}/unblock": {
      "post": {
        "operationId": "FleetController_unblock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Libera un bloqueo (CTRL-024). SOD-002: quien bloqueó no libera sin revisión.",
        "tags": [
          "fleet"
        ]
      }
    },
    "/api/v1/vehicles/{id}/service": {
      "post": {
        "operationId": "FleetController_service",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Baja / reincorporación de servicio (baja lógica, con motivo).",
        "tags": [
          "fleet"
        ]
      }
    },
    "/api/v1/drivers": {
      "get": {
        "operationId": "FleetController_listDrivers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "fleet"
        ]
      },
      "post": {
        "operationId": "FleetController_createDriver",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "fleet"
        ]
      }
    },
    "/api/v1/drivers/{id}": {
      "patch": {
        "operationId": "FleetController_updateDriver",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriverDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "fleet"
        ]
      }
    },
    "/api/v1/drivers/{id}/active": {
      "post": {
        "operationId": "FleetController_setDriverActive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverActiveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "fleet"
        ]
      }
    },
    "/api/v1/documents": {
      "get": {
        "operationId": "FleetController_listDocuments",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "fleet"
        ]
      },
      "post": {
        "operationId": "FleetController_createDocument",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Registra o renueva un documento: el anterior del mismo tipo queda «Reemplazado» (RN-009).",
        "tags": [
          "fleet"
        ]
      }
    },
    "/api/v1/maintenance-orders": {
      "get": {
        "operationId": "FleetController_listMaintenance",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "fleet"
        ]
      },
      "post": {
        "operationId": "FleetController_createMaintenance",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "fleet"
        ]
      }
    },
    "/api/v1/maintenance-orders/{id}": {
      "patch": {
        "operationId": "FleetController_updateMaintenance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMaintenanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "fleet"
        ]
      }
    },
    "/api/v1/maintenance-orders/{id}/advance": {
      "post": {
        "operationId": "FleetController_advanceMaintenance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceMaintenanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "fleet"
        ]
      }
    },
...
    "/api/v1/incidents": {
      "get": {
        "operationId": "IncidentsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "incidents"
        ]
      },
      "post": {
        "operationId": "IncidentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/v1/incidents/{id}/actions": {
      "post": {
        "operationId": "IncidentsController_addAction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/v1/incidents/{id}/advance": {
      "post": {
        "operationId": "IncidentsController_advance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "incidents"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sistema de Gestión de Transportes — API",
    "description": "DOC-E-BE · Contratos API definitivos (REST /api/v1). Errores application/problem+json (ADR-013). Autorización por permiso + alcance + política; el tenant se resuelve SIEMPRE desde la sesión.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "VehicleDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateVehicleDto": {
        "type": "object",
        "properties": {}
      },
      "ReasonDto": {
        "type": "object",
        "properties": {}
      },
      "VehicleServiceDto": {
        "type": "object",
        "properties": {}
      },
      "DriverDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateDriverDto": {
        "type": "object",
        "properties": {}
      },
      "DriverActiveDto": {
        "type": "object",
        "properties": {}
      },
      "DocumentDto": {
        "type": "object",
        "properties": {}
      },
      "MaintenanceDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateMaintenanceDto": {
        "type": "object",
        "properties": {}
      },
      "AdvanceMaintenanceDto": {
        "type": "object",
        "properties": {}
      },
...
      "IncidentDto": {
        "type": "object",
        "properties": {}
      },
      "ActionDto": {
        "type": "object",
        "properties": {}
      },
      "AdvanceDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}